long division for long pi
This is an algorithm that I was putting off for a long time, but it is
very satisfying to just have it available now! 100 digits of pi in 1
second, or 400 digits of pi in... a minute! Terrifying scaling.

Also had to switch UINF_MALLOC to a function; God only knows why the
macro was giving weird overlapping allocations, because I certainly
don't want to know.