Add support for non inverted functions
Instead of transposing the function to draw it, and bisecting it to
calculate errors from the polynomial, we just bisect it to draw it as
well, and add the option not to bisect at all.

This exposed a couple of bugs that have now been fixed. I wonder if that
was causing issues with other polynomial root finding stuff I was
doing?? Who cares, I am ditching that fixed point nonsense for simple
bisection.