continued fractions
We might be able to  use this to write our approximate polynomials as
relatively compact rational polynomials. I put this in `mod number`,
since it's kinda number theory ish in its connection to GCD and linear
combinations, (three facets of the same basic thing,) but my main
application for continued fractions is almost always going to be lossy
compression of rational numbers. Dunno if I can even structure this in a
way so that one of those three concepts would depend on the others??
It's possibly better if I just write all three independently, dunno.